+++ src/Category_Colors/CSS/Extras.php
@@ -57,7 +57,7 @@ class Extras {
 	 */
 	public static function add_list_background_css( $slug ) {
 		$css   = [];
-		$css[] = ".tribe-events-category-{$slug} h3.tribe-events-list-event-title,";
+		$css[] = ".tribe-events-category-{$slug} h3.tribe-events-list-event-title";
 		self::echo_css( $css );
 	}
 
@@ -71,7 +71,7 @@ class Extras {
 	public static function add_featured_event_link_css( $slug ) {
 		$css   = [];
 		$css[] = ".tribe-events-list .tribe-events-loop .tribe-event-featured.tribe-events-category-{$slug} h3.tribe-events-list-event-title a:hover,";
-		$css[] = "#tribe-events-content table.tribe-events-calendar .type-tribe_events.tribe-events-category-{$slug}.tribe-event-featured h3.tribe-events-month-event-title a:hover,";
+		$css[] = "#tribe-events-content table.tribe-events-calendar .type-tribe_events.tribe-events-category-{$slug}.tribe-event-featured h3.tribe-events-month-event-title a:hover";
 		self::echo_css( $css );
 	}
 
@@ -87,7 +87,7 @@ class Extras {
 		$css   = [];
 		$css[] = ".tribe-events-calendar .tribe-event-featured.tribe-events-category-{$slug},";
 		$css[] = "#tribe-events-content table.tribe-events-calendar .type-tribe_events.tribe-event-featured.tribe-events-category-{$slug},";
-		$css[] = ".tribe-grid-body div[id*='tribe-events-event-'][class*='tribe-events-category-'].tribe-events-week-hourly-single.tribe-event-featured ";
+		$css[] = ".tribe-grid-body div[id*='tribe-events-event-'][class*='tribe-events-category-'].tribe-events-week-hourly-single.tribe-event-featured";
 		$css[] = "{ border-right: 5px solid {$options['featured-event']} }";
 		self::echo_css( $css );
 	}
@@ -167,7 +167,7 @@ class Extras {
 	public static function add_deprecated_background_css( $slug ) {
 		$css   = [];
 		$css[] = "#tribe-events-content .tribe-events-category-{$slug} > .tribe-events-tooltip h4,";
-		$css[] = ".tribe-events-category-{$slug} h2,";
+		$css[] = ".tribe-events-category-{$slug} h2";
 		self::echo_css( $css );
 	}
 
@@ -181,7 +181,7 @@ class Extras {
 	public static function add_deprecated_link_css( $slug ) {
 		$css   = [];
 		$css[] = ".tribe-events-category-{$slug} h2 a,";
-		$css[] = ".tribe-events-category-{$slug} h2.tribe-events-list-event-title.entry-title a,";
+		$css[] = ".tribe-events-category-{$slug} h2.tribe-events-list-event-title.entry-title a";
 		self::echo_css( $css );
 	}
 
@@ -194,7 +194,7 @@ class Extras {
 	 */
 	public static function override_customizer( $slug ) {
 		$css   = [];
-		$css[] = ".tribe-events-shortcode .tribe-events-month table .type-tribe_events.tribe-events-category-{$slug},";
+		$css[] = ".tribe-events-shortcode .tribe-events-month table .type-tribe_events.tribe-events-category-{$slug}";
 		self::echo_css( $css );
 	}
 }