+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+* rename title to correspond to The Events Calendar standard
+* retructure CSS additions to not end with comma, added in `category.css.php`
 
 #### 7.0.6 / 2022-05-20
 * move all Pro CSS to `CSS\Pro`